pyproject.toml
[tool.poetry]
name = "calligram-documentation"
version = "0.1.0"
description = "Documentation and examples for the Calligram API."
authors = ["Nico Bautista Hobin <nico@calligram.io>"]
readme = "README.md"
packages = [{include = "calligram_documentation"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
black = "^22.10.0"
flake8 = "^6.0.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
